Move helper functions to utils
